fix(injector): [char.clone()
there is an injector bug with the mentioned fun that would cause client crashes. The fix is rather cheap and depends on us not calling char.clone() elsewhere (we don't) so its acceptable for now. will be moot in later revisions but this allows the word filter to function properly without pulling the bad bytecode out of the deob now so we will take it!